+++ htag/runners/__init__.py
@@ -19,18 +19,18 @@ try:
     from .webhttp import WebHTTP
 
 except ImportError:
-    print("You should install 'starlette' & 'uvicorn' for this htag runner")
+    print("**WARNING** Runners 'BrowserStarletteHTTP','BrowserStarletteWS' & 'WebHTTP' not availables, you should install 'starlette' & 'uvicorn' for theses htag runners")
 
 try:
     # mono instance (juste one pywebview ;-)
     from .pywebview import PyWebWiew
 except ImportError:
-    print("You should install 'pywebview' for this htag runner")
+    print("**WARNING** Runner 'PyWebWiew' not available, you should install 'pywebview' for this htag runner")
 
 try:
     # mono instance (only ! htag limitation)
     from .guyapp import GuyApp
 except ImportError:
-    print("You should install 'guy' for this htag runner")
+    print("**WARNING** Runner 'GuyApp' not available, you should install 'guy' for this htag runner")